учить загружать данные предложения
