+++ src/components/loading/loading.css
@@ -1,5 +1,6 @@
 .loading {
   width: 100%;
+  min-height: 300px;
   display: flex;
   align-items: center;
   justify-content: center;
