+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDom from 'react-dom';
 import {Provider} from 'react-redux';
 
 import App from './components/app/app';
-import {fetchOffers} from './store/data';
+import {fetchOffers} from './store/data/actions';
 import {store} from './store/store';
