+++ src/pages/favorites.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 
-import AppHeader from '../app-header/app-header';
-import AppFooter from '../app-footer/app-footer';
-import FavoritesEmpty from '../favorites-empty/favorites-empty';
-import FavoritesResult from '../favorites-result/favorites-result';
+import AppHeader from '../components/app-header/app-header';
+import AppFooter from '../components/app-footer/app-footer';
+import FavoritesEmpty from '../components/favorites-empty/favorites-empty';
+import FavoritesResult from '../components/favorites-result/favorites-result';
 
-import {OFFER_PROP_TYPES} from '../../types.js';
+import {OFFER_PROP_TYPES} from '../types.js';
 
 const Favorites = ({offers}) => {
   const favoritesOffers = offers.filter((offer) => offer.isInBookmark);
