+++ src/store/app/reducer.js
@@ -1,24 +1,6 @@
-import {City, FilterType} from '../../constants.js';
+import {City, FilterType} from '../../constants';
 import {extend} from '../../utils';
-
-export const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
-  CHANGE_FILTER: `CHANGE_FILTER`,
-};
-
-export const changeCity = (currentCity) => ({
-  type: ActionType.CHANGE_CITY,
-  payload: {
-    currentCity
-  }
-});
-
-export const changeFilter = (currentFilter) => ({
-  type: ActionType.CHANGE_FILTER,
-  payload: {
-    currentFilter,
-  }
-});
+import {ActionType} from './actions';
 
 
 const initialState = {
