+++ src/store/selectors.js
@@ -6,6 +6,8 @@ export const getOffers = (state) => state[NameSpace.DATA].offers;
 export const getCurrentCity = (state) => state[NameSpace.APP].currentCity;
 export const getCurrentFilter = (state) => state[NameSpace.APP].currentFilter;
 export const getOffersLoadingStatus = (state) => state[NameSpace.DATA].isOffersLoaded;
+export const getOfferDetailsLoadingStatus = (state) => state[NameSpace.DATA].isOfferDetailsLoaded;
+export const getOfferDetails = (state) => state[NameSpace.DATA].offerDetails;
 
 export const getFilteredOffers = createSelector(
     getOffers,
