+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: './src/actions.js',
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'public')
